minor renames: use-make-copilot-actionable
