+++ examples/next-openai/app/use-make-copilot-readable.ts
